+++ test/unit/Resources/PaymentMethod.js
@@ -5,9 +5,9 @@ const assert = chai.assert;
 const should = chai.should();
 require('dotenv').config();
 
-const Client = require('../../src/Client');
-const PaymentMethod = require('../../src/Resources/PaymentMethod');
-const Gateway = require('../../src/Resources/Gateway');
+const Client = require('../../../src/Client');
+const PaymentMethod = require('../../../src/Resources/PaymentMethod');
+const Gateway = require('../../../src/Resources/Gateway');
 
 describe('Test Gateway', () => {
